requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
nltk==3.8.1
scikit-learn==1.3.2
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
jinja2==3.1.2
openai==1.3.7
sentence-transformers==2.2.2
textstat==0.7.3
wordcloud==1.9.2
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0
reportlab==4.0.7
fpdf2==2.7.6